+++ packages/cordova/ts/app-open.ts
@@ -40,4 +40,15 @@ export class GenericAd<
   }
 }
 
-export default class AppOpenAd extends GenericAd {}
+enum AppOpenAdOrientation {
+  Portrait = 1,
+  PortraitUpsideDown = 2,
+  LandscapeRight = 3,
+  LandscapeLeft = 4,
+}
+
+export default class AppOpenAd extends GenericAd<
+  MobileAdOptions & { orientation: AppOpenAdOrientation }
+> {
+  static readonly Orientation = AppOpenAdOrientation
+}